test: 🚚 Move all queries.rs tests into dedicated integration test files
Pulls the entire 1,127-line `mod tests` block out of `src/queries.rs`
and into six topic-themed integration test files under `tests/`, so
business logic and test logic no longer share a file.

New files (37 relocated tests + the 4 already-relocated #13 variant
tests in call_site_variants.rs):

- tests/php_pattern_extraction.rs    (3 tests) — view / env / middleware
- tests/blade_pattern_extraction.rs  (5 tests) — components / directives /
                                                 attribute embeds / translations
- tests/middleware_extraction.rs     (9 tests) — every Kernel / bootstrap /
                                                 service-provider shape
- tests/feature_pattern_extraction.rs (4 tests) — Pennant call sites and
                                                  $name property capture
- tests/column_positions.rs          (15 tests) — column-position contract
                                                  for every pattern + helper
- tests/extraction_performance.rs    (1 test)  — single-pass multi-pattern
                                                 smoke test

Tests now exercise only the public crate API
(`laravel_lsp::queries::*`, `laravel_lsp::parser::*`), matching the
convention used by `tests/integration_tests.rs` and the previously
relocated `tests/call_site_variants.rs`.

Visibility change: `calculate_string_column_range` is now `pub` so its
column-math edge-case test can live alongside the other column-position
tests. The function is a pure utility (usize + &str in, Option<(usize,
usize)> out) — safe to expose.

Verification:
- cargo test: 88 lib + 64 main + 4 + 3 + 5 + 9 + 4 + 15 + 1 + 79 = 272
  passing across all targets (was 268 before the refactor)
- cargo check --tests: clean
- cargo clippy --tests: no new warnings; only pre-existing ones in
  main.rs and salsa_impl.rs
